[Component] Moved Filesystem class to its own component
