+++ src/Symfony/Bundle/SecurityBundle/Tests/Functional/WebTestCase.php
@@ -11,8 +11,8 @@
 
 namespace Symfony\Bundle\SecurityBundle\Tests\Functional;
 
-use Symfony\Component\HttpKernel\Util\Filesystem;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase as BaseWebTestCase;
+use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\HttpKernel\Kernel;
 
 class WebTestCase extends BaseWebTestCase
